feat: build and push image to nexus (#6)
* Build and push image to nexus

* fix: add pull request trigger for develop branch in GitHub Actions workflow

* fix: update dev_build job condition to trigger on all push events

* fix: reorder condition and name for environment check in CI workflow

* feat: refactor Docker setup and database initialization scripts; remove unused scripts and improve CSV handling

* fix: clean up workflow files and ensure secrets are properly defined for image build

* fix: streamline CSV push script and improve logging for successful uploads

* fix: create temporary .env file for Nexus CSV pull script and ensure cleanup

* fix: correct Docker image URL in build step to use NEXUS_DOCKER_URL

* fix: Column ID type to BIGINT + free up disk space on runner

* fix: optimize Docker image build process by adjusting variable scope and limiting concurrent uploads

* fix: automatic tagging + automatic prod build

* remove push trigger

* fix: update workflow_dispatch inputs for manual tagging in build process

* fix: remove pull_request condition for dev_build job in workflow

* fix: standardize manual_tag casing and improve version calculation handling

* fix: adjust permissions for workflow files and clean up redundant entries

* fix: remove push condition for dev_build job in workflow

* fix: remove closed pull_request type and clarify dev_build job condition

* fix: update dev_build job condition to exclude pull_request events

* fix: update dev_build job condition to trigger only on pull_request events

* fix: update manual_tag reference in prod_build job to use github.event.inputs

+++ schema/app/models/config.py
@@ -1,4 +1,4 @@
-from sqlalchemy import BigInteger, Column, Date
+from sqlalchemy import Column, Date, Integer
 from sqlalchemy.ext.declarative import declarative_base
 
 Base = declarative_base()
@@ -7,5 +7,5 @@ Base = declarative_base()
 class Config(Base):
     __tablename__ = "config"
 
-    id = Column(BigInteger, primary_key=True, autoincrement=True)
+    id = Column(Integer, primary_key=True, autoincrement=True)
     last_reset_quota_date = Column(Date)
